requirements.txt
instagrapi>=2.0.0
yt-dlp>=2024.0.0
requests>=2.31.0
pyyaml>=6.0.1
apscheduler>=3.10.0
pillow>=10.0.0